requirements.txt
streamlit>=1.32.0
numpy>=1.26.0
matplotlib>=3.8.0
Pillow>=10.1.0
pycocotools>=2.0.6
opencv-python>=4.8.0
scipy>=1.12.0
setuptools>=69.0.0
ultralytics 